requirements.txt
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.3
django-extensions==2.1.9
djangorestframework==3.9.4
factory-boy==2.12.0
Faker==1.0.7
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
